requirement.txt
streamlit
torch
torchvision
opencv-python
Pillow
numpy
streamlit-webrtc
av